+++ scripts/readme/readme_generator.py
@@ -41,7 +41,6 @@ def main(input_glob, output_files=[]):
     readme_items = sorted([j for i in globs for j in i])
 
     readme_items = local_filter(no_readme_generation_filter, readme_items)
-
     for readme in readme_items:
         readme_telemetry = Telemetry()
         workflow_name = readme.relative_to(ReadmeStepsManage.git_base_dir())