+++ src/promptflow/promptflow/executor/flow_executor.py
@@ -32,10 +32,10 @@ from promptflow.exceptions import PromptflowException
 from promptflow.executor import _input_assignment_parser
 from promptflow.executor._errors import (
     InputMappingError,
-    InvalidAggregationInput,
     NodeOutputNotFound,
     OutputReferenceBypassed,
     OutputReferenceNotExist,
+    SingleNodeValidationError,
 )
 from promptflow.executor._flow_nodes_scheduler import (
     DEFAULT_CONCURRENCY_BULK,
@@ -54,7 +54,29 @@ LINE_TIMEOUT_SEC = 600
 
 
 class FlowExecutor:
-    """This class is used to execute a single flow for different inputs."""
+    """This class is used to execute a single flow for different inputs.
+
+    :param flow: The flow to be executed.
+    :type flow: ~promptflow.contracts.flow.Flow
+    :param connections: The connections to be used for the flow.
+    :type connections: dict
+    :param run_tracker: The run tracker to be used for the flow.
+    :type run_tracker: ~promptflow._core.run_tracker.RunTracker
+    :param cache_manager: The cache manager to be used for the flow.
+    :type cache_manager: ~promptflow._core.cache_manager.AbstractCacheManager
+    :param loaded_tools: The loaded tools to be used for the flow.
+    :type loaded_tools: Mapping[str, Callable]
+    :param worker_count: The number of workers to be used for the flow. Default is 16.
+    :type worker_count: Optional[int]
+    :param raise_ex: Whether to raise exceptions or not. Default is False.
+    :type raise_ex: Optional[bool]
+    :param working_dir: The working directory to be used for the flow. Default is None.
+    :type working_dir: Optional[str]
+    :param line_timeout_sec: The line timeout in seconds to be used for the flow. Default is LINE_TIMEOUT_SEC.
+    :type line_timeout_sec: Optional[int]
+    :param flow_file: The flow file to be used for the flow. Default is None.
+    :type flow_file: Optional[Path]
+    """
 
     _DEFAULT_WORKER_COUNT = 16
 
@@ -72,6 +94,29 @@ class FlowExecutor:
         line_timeout_sec=LINE_TIMEOUT_SEC,
         flow_file=None,
     ):
+        """Initialize a FlowExecutor object.
+
+        :param flow: The Flow object to execute.
+        :type flow: ~promptflow.contracts.flow.Flow
+        :param connections: The connections between nodes in the Flow.
+        :type connections: dict
+        :param run_tracker: The RunTracker object to track the execution of the Flow.
+        :type run_tracker: ~promptflow._core.run_tracker.RunTracker
+        :param cache_manager: The AbstractCacheManager object to manage caching of results.
+        :type cache_manager: ~promptflow._core.cache_manager.AbstractCacheManager
+        :param loaded_tools: A mapping of tool names to their corresponding functions.
+        :type loaded_tools: Mapping[str, Callable]
+        :param worker_count: The number of workers to use for parallel execution of the Flow.
+        :type worker_count: int or None
+        :param raise_ex: Whether to raise an exception if an error occurs during execution.
+        :type raise_ex: bool
+        :param working_dir: The working directory to use for execution.
+        :type working_dir: str or None
+        :param line_timeout_sec: The maximum time to wait for a line of output from a node.
+        :type line_timeout_sec: int
+        :param flow_file: The path to the file containing the Flow definition.
+        :type flow_file: str or None
+        """
         # Inject OpenAI API to make sure traces and headers injection works and
         # update OpenAI API configs from environment variables.
         inject_openai_api()
@@ -133,6 +178,25 @@ class FlowExecutor:
         node_override: Optional[Dict[str, Dict[str, Any]]] = None,
         line_timeout_sec: int = LINE_TIMEOUT_SEC,
     ) -> "FlowExecutor":
+        """Create a new instance of FlowExecutor.
+
+        :param flow_file: The path to the flow file.
+        :type flow_file: Path
+        :param connections: The connections to be used for the flow.
+        :type connections: dict
+        :param working_dir: The working directory to be used for the flow. Default is None.
+        :type working_dir: Optional[str]
+        :param storage: The storage to be used for the flow. Default is None.
+        :type storage: Optional[~promptflow.storage.AbstractRunStorage]
+        :param raise_ex: Whether to raise exceptions or not. Default is True.
+        :type raise_ex: Optional[bool]
+        :param node_override: The node overrides to be used for the flow. Default is None.
+        :type node_override: Optional[Dict[str, Dict[str, Any]]]
+        :param line_timeout_sec: The line timeout in seconds to be used for the flow. Default is LINE_TIMEOUT_SEC.
+        :type line_timeout_sec: Optional[int]
+        :return: A new instance of FlowExecutor.
+        :rtype: ~promptflow.executor.flow_executor.FlowExecutor
+        """
         working_dir = Flow._resolve_working_dir(flow_file, working_dir)
         flow = Flow.from_yaml(flow_file, working_dir=working_dir)
         if node_override:
@@ -147,7 +211,7 @@ class FlowExecutor:
             flow.id, flow.name, [r.node for r in resolved_tools], inputs=flow.inputs, outputs=flow.outputs, tools=[]
         )
         # ensure_flow_valid including validation + resolve
-        # Todo: 1) split pure validation + resovle from below method 2) provide completed validation()
+        # Todo: 1) split pure validation + resolve from below method 2) provide completed validation()
         flow = FlowValidator._validate_nodes_topology(flow)
         flow.outputs = FlowValidator._ensure_outputs_valid(flow)
 
@@ -183,6 +247,23 @@ class FlowExecutor:
         working_dir: Optional[Path] = None,
         raise_ex: bool = False,
     ):
+        """Load and execute a single node from the flow.
+
+        :param flow_file: The path to the flow file.
+        :type flow_file: Path
+        :param node_name: The name of the node to be executed.
+        :type node_name: str
+        :param flow_inputs: The inputs to be used for the flow. Default is None.
+        :type flow_inputs: Optional[Mapping[str, Any]]
+        :param dependency_nodes_outputs: The outputs of the dependency nodes. Default is None.
+        :type dependency_nodes_outputs: Optional[Mapping[str, Any]
+        :param connections: The connections to be used for the flow. Default is None.
+        :type connections: Optional[dict]
+        :param working_dir: The working directory to be used for the flow. Default is None.
+        :type working_dir: Optional[str]
+        :param raise_ex: Whether to raise exceptions or not. Default is False.
+        :type raise_ex: Optional[bool]
+        """
         OperationContext.get_instance().run_mode = RunMode.SingleNode.name
         dependency_nodes_outputs = dependency_nodes_outputs or {}
 
@@ -192,9 +273,25 @@ class FlowExecutor:
             flow = Flow.deserialize(yaml.safe_load(fin))
         node = flow.get_node(node_name)
         if node is None:
-            raise ValueError(f"Node {node_name} not found in flow {flow_file}.")
+            raise SingleNodeValidationError(
+                message_format=(
+                    "Validation failed when attempting to execute the node. "
+                    "Node '{node_name}' is not found in flow '{flow_file}'. "
+                    "Please change node name or correct the flow file."
+                ),
+                node_name=node_name,
+                flow_file=flow_file,
+            )
         if not node.source or not node.type:
-            raise ValueError(f"Node {node_name} is not a valid node in flow {flow_file}.")
+            raise SingleNodeValidationError(
+                message_format=(
+                    "Validation failed when attempting to execute the node. "
+                    "Properties 'source' or 'type' are not specified for Node '{node_name}' in flow '{flow_file}'. "
+                    "Please make sure these properties are in place and try again."
+                ),
+                node_name=node_name,
+                flow_file=flow_file,
+            )
 
         flow_inputs = FlowExecutor._apply_default_value_for_input(flow.inputs, flow_inputs)
         converted_flow_inputs_for_node = FlowValidator.convert_flow_inputs_for_node(flow, node, flow_inputs)
@@ -255,16 +352,24 @@ class FlowExecutor:
 
     @staticmethod
     def update_environment_variables_with_connections(connections: dict):
+        """Update environment variables with connections.
+
+        :param connections: A dictionary containing connection information.
+        :type connections: dict
+        :return: A dictionary containing updated environment variables.
+        :rtype: dict
+        """
         from promptflow._sdk._utils import update_environment_variables_with_connections
 
         return update_environment_variables_with_connections(connections)
 
-    def convert_flow_input_types(self, inputs: dict):
-        """
-        Convert flow inputs type if existing. Ignore missing inputs.
+    def convert_flow_input_types(self, inputs: dict) -> Mapping[str, Any]:
+        """Convert the input types of the given inputs dictionary to match the expected types of the flow.
 
-        return:
-            type converted inputs
+        :param inputs: A dictionary containing the inputs to the flow.
+        :type inputs: dict
+        :return: A dictionary containing the converted inputs.
+        :rtype: Mapping[str, Any]
         """
         return FlowValidator.resolve_flow_inputs_type(self._flow, inputs)
 
@@ -273,11 +378,21 @@ class FlowExecutor:
         return {key: f"${{data.{key}}}" for key in self._flow.inputs}
 
     @property
-    def has_aggregation_node(self):
+    def has_aggregation_node(self) -> bool:
+        """Check if the flow executor has any aggregation nodes.
+
+        :return: True if the flow executor has at least one aggregation node, False otherwise.
+        :rtype: bool
+        """
         return len(self._aggregation_nodes) > 0
 
     @property
     def aggregation_nodes(self):
+        """Get the aggregation nodes of the flow executor.
+
+        :return: A list of aggregation nodes.
+        :rtype: list
+        """
         return self._aggregation_nodes
 
     @staticmethod
@@ -298,10 +413,10 @@ class FlowExecutor:
         """Collect the values from the kvs according to the indexes."""
         return {k: [v[i] for i in indexes] for k, v in kvs.items()}
 
-    def _fill_lines(self, idxes, values, nlines):
+    def _fill_lines(self, indexes, values, nlines):
         """Fill the values into the result list according to the indexes."""
         result = [None] * nlines
-        for idx, value in zip(idxes, values):
+        for idx, value in zip(indexes, values):
             result[idx] = value
         return result
 
@@ -386,7 +501,14 @@ class FlowExecutor:
             # For PromptflowException, we already do classification, so throw directly.
             raise e
         except Exception as e:
-            raise Exception(f"Failed to run aggregation nodes:\n {e}.") from e
+            error_type_and_message = f"({e.__class__.__name__}) {e}"
+            raise UnexpectedError(
+                message_format=(
+                    "Unexpected error occurred while executing the aggregated nodes. "
+                    "Please fix or contact support for assistance. The error details: {error_type_and_message}."
+                ),
+                error_type_and_message=error_type_and_message,
+            ) from e
 
     @staticmethod
     def _try_get_aggregation_input(val: InputAssignment, aggregation_inputs: dict):
@@ -398,6 +520,13 @@ class FlowExecutor:
         return InputAssignment(value=aggregation_inputs[serialized_val])
 
     def get_status_summary(self, run_id: str):
+        """Get a summary of the status of a given run.
+
+        :param run_id: The ID of the run to get the status summary for.
+        :type run_id: str
+        :return: A summary of the status of the given run.
+        :rtype: str
+        """
         return self._run_tracker.get_status_summary(run_id)
 
     def exec_aggregation(
@@ -407,45 +536,30 @@ class FlowExecutor:
         run_id=None,
         node_concurrency=DEFAULT_CONCURRENCY_FLOW,
     ) -> AggregationResult:
+        """Execute the aggregation node of the flow.
+
+        :param inputs: A mapping of input names to their values.
+        :type inputs: Mapping[str, Any]
+        :param aggregation_inputs: A mapping of aggregation input names to their values.
+        :type aggregation_inputs: Mapping[str, Any]
+        :param run_id: The ID of the current run, if any.
+        :type run_id: Optional[str]
+        :param node_concurrency: The maximum number of nodes that can be executed concurrently.
+        :type node_concurrency: int
+        :return: The result of the aggregation node.
+        :rtype: ~promptflow.executor._result.AggregationResult
+        :raises: FlowError if the inputs or aggregation_inputs are invalid.
+        """
         self._node_concurrency = node_concurrency
         aggregated_flow_inputs = dict(inputs or {})
         aggregation_inputs = dict(aggregation_inputs or {})
-        self._validate_aggregation_inputs(aggregated_flow_inputs, aggregation_inputs)
+        FlowValidator._validate_aggregation_inputs(aggregated_flow_inputs, aggregation_inputs)
         aggregated_flow_inputs = self._apply_default_value_for_aggregation_input(
             self._flow.inputs, aggregated_flow_inputs, aggregation_inputs
         )
 
         with self._run_tracker.node_log_manager:
             return self._exec_aggregation(aggregated_flow_inputs, aggregation_inputs, run_id)
-
-    @staticmethod
-    def _validate_aggregation_inputs(aggregated_flow_inputs: Mapping[str, Any], aggregation_inputs: Mapping[str, Any]):
-        """Validate the aggregation inputs according to the flow inputs."""
-        for key, value in aggregated_flow_inputs.items():
-            if key in aggregation_inputs:
-                raise InvalidAggregationInput(
-                    message_format="Input '{input_key}' appear in both flow aggregation input and aggregation input.",
-                    input_key=key,
-                )
-            if not isinstance(value, list):
-                raise InvalidAggregationInput(
-                    message_format="Flow aggregation input {input_key} should be one list.", input_key=key
-                )
-
-        for key, value in aggregation_inputs.items():
-            if not isinstance(value, list):
-                raise InvalidAggregationInput(
-                    message_format="Aggregation input {input_key} should be one list.", input_key=key
-                )
-
-        inputs_len = {key: len(value) for key, value in aggregated_flow_inputs.items()}
-        inputs_len.update({key: len(value) for key, value in aggregation_inputs.items()})
-        if len(set(inputs_len.values())) > 1:
-            raise InvalidAggregationInput(
-                message_format="Whole aggregation inputs should have the same length. "
-                "Current key length mapping are: {key_len}",
-                key_len=inputs_len,
-            )
 
     @staticmethod
     def _apply_default_value_for_aggregation_input(
@@ -505,8 +619,17 @@ class FlowExecutor:
         finally:
             remove_metric_logger(_log_metric)
 
-    def exec(self, inputs: dict, node_concurency=DEFAULT_CONCURRENCY_FLOW) -> dict:
-        self._node_concurrency = node_concurency
+    def exec(self, inputs: dict, node_concurrency=DEFAULT_CONCURRENCY_FLOW) -> dict:
+        """Executes the flow with the given inputs and returns the output.
+
+        :param inputs: A dictionary containing the input values for the flow.
+        :type inputs: dict
+        :param node_concurrency: The maximum number of nodes that can be executed concurrently.
+        :type node_concurrency: int
+        :return: A dictionary containing the output values of the flow.
+        :rtype: dict
+        """
+        self._node_concurrency = node_concurrency
         inputs = FlowExecutor._apply_default_value_for_input(self._flow.inputs, inputs)
         result = self._exec(inputs)
         #  TODO: remove this line once serving directly calling self.exec_line
@@ -545,6 +668,25 @@ class FlowExecutor:
         node_concurrency=DEFAULT_CONCURRENCY_FLOW,
         allow_generator_output: bool = False,
     ) -> LineResult:
+        """Execute a single line of the flow.
+
+        :param inputs: The input values for the line.
+        :type inputs: Mapping[str, Any]
+        :param index: The index of the line to execute.
+        :type index: Optional[int]
+        :param run_id: The ID of the flow run.
+        :type run_id: Optional[str]
+        :param variant_id: The ID of the variant to execute.
+        :type variant_id: str
+        :param validate_inputs: Whether to validate the input values.
+        :type validate_inputs: bool
+        :param node_concurrency: The maximum number of nodes that can be executed concurrently.
+        :type node_concurrency: int
+        :param allow_generator_output: Whether to allow generator output.
+        :type allow_generator_output: bool
+        :return: The result of executing the line.
+        :rtype: ~promptflow.executor._result.LineResult
+        """
         self._node_concurrency = node_concurrency
         inputs = FlowExecutor._apply_default_value_for_input(self._flow.inputs, inputs)
         # For flow run, validate inputs as default
@@ -589,13 +731,14 @@ class FlowExecutor:
         :param inputs: A list of dictionaries containing input data.
         :type inputs: List[Dict[str, Any]]
         :param run_id: Run ID.
-        :type run_id: str, optional
+        :type run_id: Optional[str]
         :param validate_inputs: Whether to validate the inputs. Defaults to True.
-        :type validate_inputs: bool, optional
-        :param raise_on_line_failure: Whether to raise an exception on line failure. Defaults to False.
-        :type raise_on_line_failure: bool, optional
+        :type validate_inputs: Optional[bool]
+        :param raise_on_line_failure: Whether to raise an exception on line failure. Defaults to False. \
+        [To be deprecated]
+        :type raise_on_line_failure: Optional[bool]
         :param node_concurrency: The node concurrency. Defaults to DEFAULT_CONCURRENCY_BULK.
-        :type node_concurrency: int, optional
+        :type node_concurrency: Optional[int]
         :return: The bulk result.
         :rtype: ~promptflow.executor.flow_executor.BulkResult
         """
@@ -634,6 +777,15 @@ class FlowExecutor:
         return updated_inputs
 
     def validate_and_apply_inputs_mapping(self, inputs, inputs_mapping) -> List[Dict[str, Any]]:
+        """Validate and apply inputs mapping for all lines in the flow.
+
+        :param inputs: The inputs to the flow.
+        :type inputs: Any
+        :param inputs_mapping: The mapping of input names to their corresponding values.
+        :type inputs_mapping: Dict[str, Any]
+        :return: A list of dictionaries containing the resolved inputs for each line in the flow.
+        :rtype: List[Dict[str, Any]]
+        """
         inputs_mapping = self._complete_inputs_mapping_by_default_value(inputs_mapping)
         resolved_inputs = self._apply_inputs_mapping_for_all_lines(inputs, inputs_mapping)
         return resolved_inputs
@@ -658,8 +810,7 @@ class FlowExecutor:
         validate_inputs: bool = False,
         allow_generator_output: bool = False,
     ) -> LineResult:
-        """
-        execute line run
+        """execute line run
 
         Args:
             inputs (Mapping): flow inputs
@@ -817,29 +968,30 @@ class FlowExecutor:
             .. code-block:: python
 
                 inputs: {
-                    "data": {"answer": 123, "question": "dummy"},
-                    "baseline": {"answer": 322},
+                    "data": {"answer": "I'm fine, thank you.", "question": "How are you?"},
+                    "baseline": {"answer": "The weather is good."},
                 }
                 inputs_mapping: {
-                    "question": "${data.question}",  # Question from the data
-                    "groundtruth": "${data.answer}",  # Answer from the data
-                    "baseline": "${baseline.answer}",  # Answer from the baseline
-                    "deployment_name": "text-davinci-003",  # literal value
+                    "question": "${data.question}",
+                    "groundtruth": "${data.answer}",
+                    "baseline": "${baseline.answer}",
+                    "deployment_name": "literal_value",
                 }
 
                 Returns: {
-                    "question": "dummy",
-                    "groundtruth": 123,
-                    "baseline": 322,
-                    "deployment_name": "text-davinci-003",
+                    "question": "How are you?",
+                    "groundtruth": "I'm fine, thank you."
+                    "baseline": "The weather is good.",
+                    "deployment_name": "literal_value",
                 }
 
-        :param inputs: Inputs for the flow.
+        :param inputs: A mapping of input keys to their corresponding values.
         :type inputs: Mapping[str, Mapping[str, Any]]
-        :param inputs_mapping: Inputs mapping for the flow.
+        :param inputs_mapping: A mapping of input keys to their corresponding mapping expressions.
         :type inputs_mapping: Mapping[str, str]
-        :return: Processed inputs for the flow.
+        :return: A dictionary of input keys to their corresponding mapped values.
         :rtype: Dict[str, Any]
+        :raises InputMappingError: If any of the input mapping relations are not found in the inputs.
         """
         import re
 
@@ -855,7 +1007,7 @@ class FlowExecutor:
             match = re.search(r"^\${([^{}]+)}$", map_value)
             if match is not None:
                 pattern = match.group(1)
-                # Could also try each paire of key value from inputs to match the pattern.
+                # Could also try each pair of key value from inputs to match the pattern.
                 # But split pattern by '.' is one deterministic way.
                 # So, give key with less '.' higher priority.
                 splitted_str = pattern.split(".")
@@ -939,12 +1091,12 @@ class FlowExecutor:
                             tmp_dict[index] = {}
                         tmp_dict[index][input_key] = one_line_item
         result = []
-        for line, valus_for_one_line in tmp_dict.items():
+        for line, values_for_one_line in tmp_dict.items():
             # Missing input is not acceptable line.
-            if len(valus_for_one_line) != len(input_dict):
+            if len(values_for_one_line) != len(input_dict):
                 continue
-            valus_for_one_line[LINE_NUMBER_KEY] = line
-            result.append(valus_for_one_line)
+            values_for_one_line[LINE_NUMBER_KEY] = line
+            result.append(values_for_one_line)
         return result
 
     @staticmethod
@@ -1009,10 +1161,13 @@ class FlowExecutor:
         return result
 
     def enable_streaming_for_llm_flow(self, stream_required: Callable[[], bool]):
-        """Enable the LLM node that is connected to output to return streaming results controlled by stream_required.
+        """Enable the LLM node that is connected to output to return streaming results controlled by `stream_required`.
 
         If the stream_required callback returns True, the LLM node will return a generator of strings.
         Otherwise, the LLM node will return a string.
+        :param stream_required: A callback that takes no arguments and returns a boolean value indicating whether
+        streaming results should be enabled for the LLM node.
+        :type stream_required: Callable[[], bool]
         """
         for node in self._flow.nodes:
             if (
@@ -1031,6 +1186,8 @@ class FlowExecutor:
 
         This method adds a wrapper to each node in the flow
         to consume the streaming outputs and merge them into a string for executor usage.
+
+        :return: None
         """
         for node in self._flow.nodes:
             self._tools_manager.wrap_tool(node.name, wrapper=_ensure_node_result_is_serializable)
@@ -1062,7 +1219,12 @@ def _inject_stream_options(should_stream: Callable[[], bool]):
 
 
 def enable_streaming_for_llm_tool(f):
-    """Enable the stream mode for LLM tools that supports it.
+    """Enable the stream mode for LLM tools that support it.
+
+    :param f: The function to wrap.
+    :type f: function
+    :return: The wrapped function.
+    :rtype: function
 
     AzureOpenAI.completion and AzureOpenAI.chat tools support both stream and non-stream mode.
     The stream mode is turned off by default. Use this wrapper to turn it on.